Use bytes as unit in cache and extend tests

+++ src/util/files_mock.go
@@ -12,7 +12,7 @@ type MockFilesystem struct {
 	MkdirAllFunc        func(path string) error
 	CreateTempFunc      func(dir, pattern string) (*os.File, error)
 	DirSizeInBytesFunc  func(path string) (error, int64)
-	FindLargestFileFunc func(path string) (error, float64, string)
+	FindLargestFileFunc func(path string) (error, int64, string)
 }
 
 func (m *MockFilesystem) Exists(path string) bool {
@@ -43,6 +43,6 @@ func (m *MockFilesystem) DirSizeInBytes(path string) (error, int64) {
 	return m.DirSizeInBytesFunc(path)
 }
 
-func (m *MockFilesystem) FindLargestFile(path string) (error, float64, string) {
+func (m *MockFilesystem) FindLargestFile(path string) (error, int64, string) {
 	return m.FindLargestFileFunc(path)
 }
